Fix bugs

+++ src/main/java/com/projects/oleksii/leheza/cashtruck/dto/view/CategoryInfoDto.java
@@ -15,6 +15,6 @@ public class CategoryInfoDto {
     @NotBlank
     private String name;
     private double fullCategoryTransactionSum;
-    private int categoryPercentage;
+    private double categoryPercentage;
     private int currencyCode;
 }
